split node into own module
