+++ trees/src/main.rs
@@ -1,9 +1,10 @@
 mod rb_tree;
-mod avl_tree;
+//mod avl_tree;
 mod tree;
+mod node;
 use crate::rb_tree::*;
 use crate::tree::*;
-use crate::avl_tree::*;
+//use crate::avl_tree::*;
 
 fn main() {
     let mut tree = RedBlackTree::new();
@@ -31,5 +32,5 @@ fn main() {
     tree2.print_inorder();
     tree2.print_structure();
 
-    let mut avl:AvlTree<i32> = AvlTree::new();
+   // let mut avl:AvlTree<i32> = AvlTree::new();
 }
