+++ trees/src/rb_tree.rs
@@ -1,4 +1,5 @@
 use crate::tree::*;
+use crate::node::*;
 use std::borrow::{Borrow, BorrowMut};
 use std::cell::RefCell;
 use std::rc::Rc;
